Refactored: removed stream redundancy
