+++ index.html
@@ -2,9 +2,9 @@
 <html lang="en">
 <head>
     <meta charset="UTF-8">
-    <title>streaming pages</title>
+    <title>served from node http</title>
 </head>
 <body>
-    <h1>served from node http</h1>
+    <h1>hello, world!</h1>
 </body>
 </html>
